[SLH] AArch64: refactor CFG instrumentation logic
When MisspeculatingTaintReg is not available, do not compute temporary
registers as the functions that should use them later are no-op when
falling back to full barriers. This makes it more obvious which
instrumentation is emitted when using full barriers across the whole
function and makes it possible to always assume MisspeculatingTaintReg
is available when computing temporary registers.

Note that this commit is not strictly NFC because call instructions
were not instrumented when TmpRegisterNotAvailableEverywhere is true.
This seems to be incorrect w.r.t. preventing mis-speculation
coming from the callee, so now this code is inserted unconditionally.