[fixup] Switch tests to IA key
IA is the key used by default in mainline LLVM.

There are possibly tests in mainline checking the same behavior, but
keeping the test for now to make sure the expected behavior is the same
both in mainline and this patch. The test may further be
shrinked/removed when upstreaming the patch, though.