[fixup] Remove reimplementation of already existing features
In 7924c7d75ae0, a "ptrauth-returns" function attribute is used to choose
which functions to harden. In mainline LLVM, a global option is used instead
of a per-function one. It should be possible to trivially implement this
check in `AArch64FunctionInfo::shouldSignReturnAddress` if needed.

Aside from deciding which function to harden and "loop over a collection of
spilled registers with at most one LR" vs. "check for spilled LR in a separate
unction" patterns of code, the removed parts are quite similar to already
upstreamed code.