[fixup] Ported LR auth check on tail-call to mainline LLVM
Switched from "ptrauth-*" function attributes to
AArch64FunctionInfo.shouldSignReturnAddress(), added
aarch64-auth-check-lr-tc command line option.

FIXME: Current implementation assumes address bits 62:63 to be the
top-most bits of PAC so it does not support TBI. The original
implementation did not support it either and was explicitly disabled on
non-Darwin targets (so TBI was known to be disabled).