[AArch64][PAC] Implement code generation for @llvm.ptrauth.resign
Introduce PAUTH_RESIGN pseudo instruction that is similar to PAUTH_AUTH
but has two sets of operands expressing signing schemas: the "old" schema
used to authenticate the input pointer and the "new" schema used for
signing.

PAUTH_RESIGN is expanded in the same way as PAUTH_AUTH immediately followed
by signing operation using the "new" schema. Between authentication and
signing operations, the same check is inserted that would be used by
PAUTH_AUTH, with method chosen based on the key used for authentication.