[clang][AArch64][PAC] Fix vtable pointer discrimination for RTTI types
For RTTI types, like `__class_type_info`, we do not have their full definition
when emitting AUTH relocations matching their vtable pointers. So, we were
unable to calculate type-dependent discriminator properly previously.

To fix that, an additional command-line option
`-fptrauth-cxx-type-info-vtable-discriminator` is proposed. Three values
are allowed:

- `zero`: use constant discriminator of 0;
- `type`: use type-dependent discriminator calculated with only hard-coded
  mangled type name in mind; may be suitable for implementations where RTTI
  classes are not inherited from other ones, for example, for libstdc++;
- `libcxx`: use constant discriminator matching base interface class
  type `__shim_type_info` which is used for RTTI types in libcxx.

The option is set to `libcxx` by default when building for ELF platforms with
`-mbranch-protection=pauthabi`. The vtable type info discrimination is
enabled correspondingly.

Address discrimination is enabled the same way as for the reset of vtable
pointers - via `-fptrauth-vtable-pointer-address-discrimination`. It is
enabled by default with `-mbranch-protection=pauthabi`.