[FIXME][lld][AArch64][PAC] Fix unreadable chars issue in `Twine` error message
For some reason, `Twine` error message obtained via concatenation of
string literals and `StringRef`'s (or `std::string`'s - the result was
the same) contained unreadable characters, which caused test failure.

The `StringRef`'s containins file names are **not** corrupted: when
printing them separately, they are OK, and calling `.str()` on them also
gives a correct `std::string` object. Moreover, a `Twine` constructed
from a file name only (without concatenation with other stuff) also
contains correct string without unreadable characters.

The issue is not related to thread-safety: using a single thread does
not resolve it.

The issue is present when both g++ and clang++ are used to compile llvm
and lld. The output is a bit different though, but if the issue is related to
reading corrupt data from somewhere, different output is what we expect.