[AArch64][PAC] Emit a fatal error when ptrauth target feature is missing
Previously, in `expandPtrAuthPseudo` we expanded `MOVaddrPAC`,
`LOADgotPAC` and `LOADauthptrgot` pseudo-instructions without taking
presense of PAuth subtarget feature into account. In case of
`LOADauthptrgot`, it resulted in undesired so-called `$auth_ptr$` stub and a
corresponding AUTH relocation. In case `MOVaddrPAC` and `LOADgotPAC`, it
resulted in pac-specific instructions (e.g. `paciza`) emitted which was
only caught via assertion during machine instructions verification (and
not caught at all when assertions are disabled).

This patch makes us emit a fatal error and fail fast in such cases.