[AArch64][PAuth] Support BLRA* instructions in SLS Hardening pass
Support hardening authenticated variants of BLR instruction against
Straight Line Speculation.

The existing implementation generates __llvm_slsblr_thunk_xN thunk
functions for each N=0..31, except 16, 17 and 30, even if unused.
As explained in the comments, this would require generating about 1800
thunks (most of which are likely to be unused) for BLRA* instructions,
as BLRAA and BLRAB accept two register operands.

This commit introduces four more kinds of thunks in addition to
__llvm_slsblr_thunk_xN and makes SLS Hardening machinery keep track
which ones are actually called instead of generating all of them.