[PAC][clang] Error if PAuth language features are used on unsupported CPU
Check for several `-fptrauth-*` flags corresponding to language features that
require CPU with `FEAT_PAuth` support. If such unsupported feature is
requested, make Clang frontend print an error and exit early.

Also ignore some flags intended to be used only with `-fptrauth-calls`
if this condition is not met. List of such `-fptrauth-*` flags:
- `vtable-pointer-address-discrimination`;
- `vtable-pointer-type-discrimination`;
- `init-fini`.