[AArch64][PAC] Reuse AddrDisc when computing blend, if possible
As function calls do not usually preserve X16 and X17, it is beneficial
to allow AddrDisc operand of B(L)RA instruction to reside in these
registers. This saves one MOV instruction by updating AddrDisc in-place
if it ends up allocated to X16 or X17 and integer discriminator is non-zero.